SQL/requirements.txt
fastapi
uvicorn
numpy
mysql-connector-python
pydantic
dotenv
joblib
tensorflow